requirements.txt
flask==3.0.0
python-docx==1.1.0
requests==2.31.0
jsonschema==4.20.0
gunicorn==21.2.0
pytest==7.4.3
python-dotenv==1.0.0
werkzeug==3.0.1
flask-cors==4.0.0
